requirements.txt
django==4.1.3
ipython
django-import-export
